Adds a submodule for the default ckan theme/schema etc.
To be able to work in defining a new schema, it is necessary to disable
the valdiated_dict in the solr schema and re-index the content.  If you
don't do this then the schema isn't used during package_show because of
the validated cache.

Requires fetching the git submodule in src/ and also re-running
```vagrant provision```.