Add a user_auth endpoint for Alpha.
During Alpha we wish to be able to authenticate users without needing a
local model, and so here is an API endpoint, callable only by a sysadmin
which will return all of the user's detail on correct submission of
email and password.

The resulting dictionary also contains a list of the organisations that
the user is in.

One stumbling block is that we currently allow emails to be re-used by
different users.  For now we are just using the first such match, but
over time (as per #4) we need to resolve this (if any actually exist).