Show an upload option if organogram appears in title
As we need to explore options for allowing file uploads when users are
publishing organogram datasets, we should only show the option when
we can determine it is such a dataset.

Initially, we have no mechanism for starting the process of creating an
organogram dataset, and so we will currently rely on the word 'organogram'
appearing in the dataset title.  This should change once we have a more
accurate way of initiating the process.