Remove the "Clear" button for harvest sources
This doesn't behave the way users will expect given other changes
to our CKAN, so we're just going to remove it.

This is the minimal amount of template I could copy over from the
harvest plugin, unfortunately there's no more granular `block` hooks to
override.

This means we should keep an eye on changes to this section of the
template when upgrading the harvest plugin.