Remove inheritance from CKAN for testing
There were two test configurations: `test.ini` and `test-jenkins.ini`.
These both inherited another configuration from core CKAN, which is
located in a different location for each test environment.

Changes to the test configuration had to be made in both files, which
resulted in duplication of effort and possibility for things to be
missed.

This change removes the inheritance from core CKAN and removes
duplication from `test-jenkins.ini`, which now inherits from the main
test configuration.