Add GC-Data-Catalogue-csv-schema
This is so that publishers can validate their datasets against it.

Schemas are used to validate datasets on data.gov.uk against a
specification.
The geospatial commission is doing some work at the moment that they
want to make sure complies to a schema that hasn't been used on
data.gov.uk before.

Trello: https://trello.com/c/KkaFHYpE/749-add-a-new-schema-to-datagovuk