Disallow indexing by search engine bots
The way we use CKAN is purely as an administration interface to
manage the datasets which are rendered on data.gov.uk using a
separate 'Find' application.

Therefore we don't want any robots to be indexing any of the
pages in CKAN as they will only be accessible to people who have
logged in.

Also we hope this will reduce the 5xx rate on CKAN as bots are
frequently requesting a page which is raising a 5xx error and
we'd rather stop them accessing this page rather than fixing the
issue in CKAN itself.