Fix logging and refine script
The script will find filenames in S3 that don't end in `senior.csv`,
'junior.csv' or include `-posts-`, and replace them with the
correct file structure.

It also replaces the correponding `url` in the database. This ensures
the url shown for its dataset in Find and CKAN matches, and shows
the organogram visualisation.