Add solr reindex script which targets a date
As Publish and CSW use the CKAN API which in turn uses SOLR to extract data, we need to ensure that SOLR is reindexed if we make ad hoc changes to the database. As long as the metdata_modified date is later than the one set in the Publish DB, the dataset will be updated.