Update styling of active navigation items
During the upgrade from 2.8 to 2.9, the links between sections in the ckan UI changed to include a trailing slash on the links. This broke the styling we currently have for links that are specified via the href. This feature is useful to communicate to users where they are within the ckan service.

This level of specificity in our CSS is less than ideal, however we aren't in full control of this setup as the navigation is handled within ckan (not this extension).

I removed the additional calls to li.active a as I don't envision a situation where we would ever have 2 "active" li's at any one time. We also don't have a site usage page anymore that's accessible from the header navigation so I've removed this as well.